+++ crates/data-dict/src/assert_expr.rs
@@ -161,12 +161,7 @@ pub enum CmpOp {
     Ge,
 }
 
-#[derive(Debug, Clone)]
-pub struct ParseError {
-    pub message: String,
-    /// Byte offset of the failing token (or end-of-string) within the input.
-    pub at: usize,
-}
+pub use crate::expr_lex::ParseError;
 
 impl AssertExpr {
     pub fn parse(input: &str) -> Result<AssertExpr, ParseError> {
@@ -470,7 +465,7 @@ impl<'a> Parser<'a> {
             }
             Some(b'\'') => self.parse_string(),
             Some(b'`') => {
-                let name = self.parse_quoted_name()?;
+                let name = crate::expr_lex::parse_quoted_name(self.src, &mut self.pos)?;
                 Ok(self.node(ExprKind::Column(name), start))
             }
             Some(b) if b.is_ascii_digit() => self.parse_number(),
@@ -499,7 +494,7 @@ impl<'a> Parser<'a> {
                 }
                 Some(_) => {
                     let ch_start = self.pos;
-                    self.advance_char();
+                    crate::expr_lex::advance_char(self.src, &mut self.pos);
                     value.push_str(
                         std::str::from_utf8(&self.src[ch_start..self.pos])
                             .expect("input is valid utf-8"),
@@ -512,63 +507,6 @@ impl<'a> Parser<'a> {
             start,
             end: self.pos,
         })
-    }
-
-    /// Parse a backtick-quoted column name, leaving `self.pos` after the
-    /// closing backtick. Errors point at the opening backtick, the token the
-    /// reader has to fix.
-    fn parse_quoted_name(&mut self) -> Result<String, ParseError> {
-        let start = self.pos;
-        debug_assert_eq!(self.peek(), Some(b'`'));
-        self.pos += 1;
-        let mut name = String::new();
-        loop {
-            match self.peek() {
-                None => {
-                    return Err(ParseError {
-                        message: "unterminated quoted name".into(),
-                        at: start,
-                    });
-                }
-                Some(b'`') => {
-                    // A doubled backtick is a literal backtick; a lone one ends it.
-                    if self.src.get(self.pos + 1) == Some(&b'`') {
-                        name.push('`');
-                        self.pos += 2;
-                    } else {
-                        self.pos += 1;
-                        break;
-                    }
-                }
-                Some(_) => {
-                    let ch_start = self.pos;
-                    self.advance_char();
-                    name.push_str(
-                        std::str::from_utf8(&self.src[ch_start..self.pos])
-                            .expect("input is valid utf-8"),
-                    );
-                }
-            }
-        }
-        if name.is_empty() {
-            return Err(ParseError {
-                message: "empty quoted name".into(),
-                at: start,
-            });
-        }
-        Ok(name)
-    }
-
-    fn advance_char(&mut self) {
-        // Step over one whole UTF-8 code point.
-        self.pos += 1;
-        while let Some(&b) = self.src.get(self.pos) {
-            if b & 0xC0 == 0x80 {
-                self.pos += 1;
-            } else {
-                break;
-            }
-        }
     }
 
     fn parse_number(&mut self) -> Result<Expr, ParseError> {
@@ -714,7 +652,7 @@ impl<'a> Parser<'a> {
                     self.skip_ws();
                     let n_start = self.pos;
                     let name = match self.peek() {
-                        Some(b'`') => self.parse_quoted_name()?,
+                        Some(b'`') => crate::expr_lex::parse_quoted_name(self.src, &mut self.pos)?,
                         Some(b) if b.is_ascii_alphabetic() || b == b'_' => self.read_word(),
                         _ => return Err(self.err("expected a column name")),
                     };
